Remove PromptSelectField component
