--- vscode/webviews/components/promptSelectField/fixtures.ts
@@ -1,52 +1,0 @@
-import type { Prompt } from '@sourcegraph/cody-shared'
-
-export const FIXTURE_PROMPTS: Prompt[] = [
-    {
-        id: '1',
-        name: 'TypeScript Vitest Test',
-        nameWithOwner: 'alice/TypeScript Vitest Test',
-        owner: { namespaceName: 'alice', displayName: 'Alice Zhao' },
-        description: 'Generate unit tests for a given function',
-        draft: false,
-        definition: { text: 'Generate unit tests for vitest' },
-        url: 'https://example.com',
-    },
-    {
-        id: '2',
-        name: 'Review OpenCtx Provider',
-        nameWithOwner: 'alice/Review OpenCtx Provider',
-        owner: { namespaceName: 'alice', displayName: 'Alice Zhao' },
-        description: 'Suggest improvements for an OpenCtx provider',
-        draft: true,
-        definition: { text: 'Review the following OpenCtx provider code' },
-        url: 'https://example.com',
-    },
-    {
-        id: '3',
-        name: 'Generate JUnit Integration Test',
-        nameWithOwner: 'myorg/Generate JUnit Integration Test',
-        owner: { namespaceName: 'myorg', displayName: 'My Org' },
-        draft: false,
-        definition: { text: 'Generate a JUnit integration test' },
-        url: 'https://example.com',
-    },
-    {
-        id: '4',
-        name: 'Fix Bazel Build File',
-        nameWithOwner: 'myorg/Fix Bazel Build File',
-        owner: { namespaceName: 'myorg', displayName: 'My Org' },
-        draft: false,
-        definition: { text: 'Fix common issues in this Bazel BUILD file' },
-        url: 'https://example.com',
-    },
-    {
-        id: '5',
-        name: 'Convert from React Class to Function Component',
-        nameWithOwner: 'abc-corp/Convert from React Class to Function Component',
-        owner: { namespaceName: 'abc-corp', displayName: 'ABC Corp' },
-        description: 'Convert from a React class component to a function component',
-        draft: false,
-        definition: { text: 'Convert from a React class component to a function component' },
-        url: 'https://example.com',
-    },
-]
